Add timeout check for configuration backup
The most time consuming operation in configuration backup is generating
the tar ball. In "kimchi.utils", the run_command() function already
supports setting timeout, so in this patch we just set the timeout for
running the tar command. Once the caller catches the TimeoutExpired
exception, it knows the files are too large for a configuration backup,
and informs the user.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
(cherry picked from commit 3cc5181bcdb2eea03e7589ea462c2b74de0dee10)