Fix Firmware level information of BMC-based machines.
The output of lsmcode command on BMC-based machines is extremely different than
its output on FSP-based machines. Because of that, the information provided in
Administration tab regarding the current Firmware level is not being pretty
formatted in BMC-based machines.

This patch parses the output of lsmcode command to get the correct information
about the Firmware level installed.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>