Improve the way "Updates" tab is loaded
The "Updates" tab was requesting information about the host to get
loaded which is not needed.
Also, there were some HTML elements without ID or with wrong reference to an ID
(in those cases the widget needs to know the parent (data-parent) and the content (href and aria-controls))
This patch also removes unused CSS.

All those issues were identified while investigating the cause of the
problem reported on #122.